backend/requirements.txt
flask==3.1.0
flask-cors==3.0.10
flask-sqlalchemy==3.1.1
psycopg2-binary==2.9.10
python-dotenv==1.0.0
google-cloud-aiplatform==1.83.0
google-genai==1.5.0
